pyproject.toml
[tool.poetry]
name = "pdfgrabba"
version = "0.1.0"
description = "Semi-interactive CLI for downloading academic PDFs from a BibTeX file"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/yourusername/pdfgrabba"
packages = [{ include = "pdfgrabba" }]

[tool.poetry.dependencies]
python = "^3.10"
bibtexparser = "^1.4"
requests = "^2.31"
undetected-chromedriver = "^3.5"
selenium = "^4.0"
setuptools = "*"
pyyaml = "^6.0"
pydantic = "^2.0"

[tool.poetry.group.dev.dependencies]
black = "^24.0"
ruff = "^0.4"

[tool.poetry.scripts]
pdfgrabba = "pdfgrabba.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
